Reorganize matcher tests by module
